+++ controllers/LoaiSanPhamController.php
@@ -6,9 +6,13 @@ class LoaiSanPhamController
     public function index()
     {
         if (isset($_SESSION['role']) && $_SESSION['role'] != 4) {
-            $LoaiTruyenDAO = new LoaiTruyenDAO();
-            $list = $LoaiTruyenDAO->show();
-            include_once "views/danhmuc/admin/list.php";
+            if (isset($_SESSION['chuyen'])) {
+                header('location: index.php?controller=trangChu');
+            } else {
+                $LoaiTruyenDAO = new LoaiTruyenDAO();
+                $list = $LoaiTruyenDAO->show();
+                include_once "views/danhmuc/admin/list.php";
+            }
         } else {
             header('location: index.php?controller=trangChu');
         }
@@ -17,15 +21,19 @@ class LoaiSanPhamController
     public function add()
     {
         if (isset($_SESSION['role']) && $_SESSION['role'] != 4) {
-            if (isset($_POST['ten'])) {
-                $LoaiTruyenDAO = new LoaiTruyenDAO();
-                $LoaiTruyenDAO->add($_POST['ten']);
-                $list = $LoaiTruyenDAO->show();
-                $_SESSION['error'] = 'thêm mới thành công';
-                header('location: index.php?controller=loaisanpham');
-                exit();
+            if (isset($_SESSION['chuyen'])) {
+                header('location: index.php?controller=trangChu');
             } else {
-                include_once('views/danhmuc/admin/add.php');
+                if (isset($_POST['ten'])) {
+                    $LoaiTruyenDAO = new LoaiTruyenDAO();
+                    $LoaiTruyenDAO->add($_POST['ten']);
+                    $list = $LoaiTruyenDAO->show();
+                    $_SESSION['error'] = 'thêm mới thành công';
+                    header('location: index.php?controller=loaisanpham');
+                    exit();
+                } else {
+                    include_once('views/danhmuc/admin/add.php');
+                }
             }
         } else {
             header('location: index.php?controller=trangChu');
@@ -35,12 +43,16 @@ class LoaiSanPhamController
     public function remove()
     {
         if (isset($_SESSION['role']) && $_SESSION['role'] != 4) {
-            $LoaiTruyenDAO = new LoaiTruyenDAO();
-            $LoaiTruyenDAO->remove($_GET['id']);
-            $list = $LoaiTruyenDAO->show();
-            $_SESSION['error'] = 'Xoá thành công';
-            header('location: index.php?controller=loaisanpham');
-            exit();
+            if (isset($_SESSION['chuyen'])) {
+                header('location: index.php?controller=trangChu');
+            } else {
+                $LoaiTruyenDAO = new LoaiTruyenDAO();
+                $LoaiTruyenDAO->remove($_GET['id']);
+                $list = $LoaiTruyenDAO->show();
+                $_SESSION['error'] = 'Xoá thành công';
+                header('location: index.php?controller=loaisanpham');
+                exit();
+            }
         } else {
             header('location: index.php?controller=trangChu');
         }
@@ -49,17 +61,21 @@ class LoaiSanPhamController
     public function update()
     {
         if (isset($_SESSION['role']) && $_SESSION['role'] != 4) {
-            if (isset($_POST['ten'])) {
-                $LoaiTruyenDAO = new LoaiTruyenDAO();
-                $LoaiTruyenDAO->update($_POST['id'], $_POST['ten'], $_POST['trang_thai']);
-                $list = $LoaiTruyenDAO->show();
-                $_SESSION['error'] = 'Sửa thông tin thành công';
-                header('location: index.php?controller=loaisanpham');
-                exit();
+            if (isset($_SESSION['chuyen'])) {
+                header('location: index.php?controller=trangChu');
             } else {
-                $LoaiTruyenDAO = new LoaiTruyenDAO();
-                $list = $LoaiTruyenDAO->showOne($_GET['id']);
-                include_once "views/danhmuc/admin/fix.php";
+                if (isset($_POST['ten'])) {
+                    $LoaiTruyenDAO = new LoaiTruyenDAO();
+                    $LoaiTruyenDAO->update($_POST['id'], $_POST['ten'], $_POST['trang_thai']);
+                    $list = $LoaiTruyenDAO->show();
+                    $_SESSION['error'] = 'Sửa thông tin thành công';
+                    header('location: index.php?controller=loaisanpham');
+                    exit();
+                } else {
+                    $LoaiTruyenDAO = new LoaiTruyenDAO();
+                    $list = $LoaiTruyenDAO->showOne($_GET['id']);
+                    include_once "views/danhmuc/admin/fix.php";
+                }
             }
         } else {
             header('location: index.php?controller=trangChu');
