+++ controllers/NhaPhatHanhController.php
@@ -6,9 +6,13 @@ class NhaPhatHanhController
     public function index()
     {
         if (isset($_SESSION['role']) && $_SESSION['role'] != 4) {
-            $NhaPhatHanhDAO = new NhaPhatHanhDAO();
-            $list = $NhaPhatHanhDAO->show();
-            include_once "views/nhaphathanh/admin/list.php";
+            if (isset($_SESSION['chuyen'])) {
+                header('location: index.php?controller=trangChu');
+            } else {
+                $NhaPhatHanhDAO = new NhaPhatHanhDAO();
+                $list = $NhaPhatHanhDAO->show();
+                include_once "views/nhaphathanh/admin/list.php";
+            }
         } else {
             header('location: index.php?controller=trangChu');
         }
@@ -17,15 +21,19 @@ class NhaPhatHanhController
     public function add()
     {
         if (isset($_SESSION['role']) && $_SESSION['role'] != 4) {
-            if (isset($_POST['ten'])) {
-                $NhaPhatHanhDAO = new NhaPhatHanhDAO();
-                $NhaPhatHanhDAO->add($_POST['ten']);
-                $list = $NhaPhatHanhDAO->show();
-                $_SESSION['error'] = 'thêm mới thành công';
-                header('location: index.php?controller=nhaPhatHanh');
-                exit();
+            if (isset($_SESSION['chuyen'])) {
+                header('location: index.php?controller=trangChu');
             } else {
-                include_once('views/nhaphathanh/admin/add.php');
+                if (isset($_POST['ten'])) {
+                    $NhaPhatHanhDAO = new NhaPhatHanhDAO();
+                    $NhaPhatHanhDAO->add($_POST['ten']);
+                    $list = $NhaPhatHanhDAO->show();
+                    $_SESSION['error'] = 'thêm mới thành công';
+                    header('location: index.php?controller=nhaPhatHanh');
+                    exit();
+                } else {
+                    include_once('views/nhaphathanh/admin/add.php');
+                }
             }
         } else {
             header('location: index.php?controller=trangChu');
@@ -35,12 +43,16 @@ class NhaPhatHanhController
     public function remove()
     {
         if (isset($_SESSION['role']) && $_SESSION['role'] != 4) {
-            $NhaPhatHanhDAO = new NhaPhatHanhDAO();
-            $NhaPhatHanhDAO->remove($_GET['id']);
-            $list = $NhaPhatHanhDAO->show();
-            $_SESSION['error'] = 'Xoá thành công';
-            header('location: index.php?controller=nhaPhatHanh');
-            exit();
+            if (isset($_SESSION['chuyen'])) {
+                header('location: index.php?controller=trangChu');
+            } else {
+                $NhaPhatHanhDAO = new NhaPhatHanhDAO();
+                $NhaPhatHanhDAO->remove($_GET['id']);
+                $list = $NhaPhatHanhDAO->show();
+                $_SESSION['error'] = 'Xoá thành công';
+                header('location: index.php?controller=nhaPhatHanh');
+                exit();
+            }
         } else {
             header('location: index.php?controller=trangChu');
         }
@@ -49,17 +61,21 @@ class NhaPhatHanhController
     public function update()
     {
         if (isset($_SESSION['role']) && $_SESSION['role'] != 4) {
-            if (isset($_POST['ten'])) {
-                $NhaPhatHanhDAO = new NhaPhatHanhDAO();
-                $NhaPhatHanhDAO->update($_POST['id'], $_POST['ten'], $_POST['trang_thai']);
-                $list = $NhaPhatHanhDAO->show();
-                $_SESSION['error'] = 'Sửa thông tin thành công';
-                header('location: index.php?controller=nhaPhatHanh');
-                exit();
+            if (isset($_SESSION['chuyen'])) {
+                header('location: index.php?controller=trangChu');
             } else {
-                $NhaPhatHanhDAO = new NhaPhatHanhDAO();
-                $list = $NhaPhatHanhDAO->showOne($_GET['id']);
-                include_once "views/nhaphathanh/admin/fix.php";
+                if (isset($_POST['ten'])) {
+                    $NhaPhatHanhDAO = new NhaPhatHanhDAO();
+                    $NhaPhatHanhDAO->update($_POST['id'], $_POST['ten'], $_POST['trang_thai']);
+                    $list = $NhaPhatHanhDAO->show();
+                    $_SESSION['error'] = 'Sửa thông tin thành công';
+                    header('location: index.php?controller=nhaPhatHanh');
+                    exit();
+                } else {
+                    $NhaPhatHanhDAO = new NhaPhatHanhDAO();
+                    $list = $NhaPhatHanhDAO->showOne($_GET['id']);
+                    include_once "views/nhaphathanh/admin/fix.php";
+                }
             }
         } else {
             header('location: index.php?controller=trangChu');
